Added parameter units
When parameters are set up, the units are read from the parnames.txt file
and added to the model object. For the likelihood parameters, they are
read from the par_likeli.txt file. Idea: they can be used to adapt the
parameter values to the desired time units. E.g. calculation was with
6h-timesteps (and so are parameters) but output is wished in 1h-time
units.