Distinguish different outnames.txt files
This makes it possible to run different model structures at the same time.
The names of the output (outnames) are stored in the sudriv object when
model.setup is run. They are acessed by default through the
result_index_var function in run.model. Backward compatibility with
previous runs of result_index_var (e.g. in plotting) is kept.